+++ .gitignore
@@ -62,3 +62,7 @@ pom.xml.tag
 pom.xml.releaseBackup
 pom.xml.versionsBackup
 release.properties
+
+
+# Asciidoctor
+.asciidoctor
