+++ asciidoc/asciidoc-light/deployment/src/main/java/io/quarkiverse/qute/web/asciidoclight/deployment/QuteWebAsciidocProcessor.java
@@ -1,0 +1,38 @@
+package io.quarkiverse.qute.web.asciidoclight.deployment;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import io.quarkiverse.qute.web.asciidoclight.runtime.AsciidocLightSectionHelperFactory;
+import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
+import io.quarkus.deployment.annotations.BuildProducer;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.CombinedIndexBuildItem;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
+
+class QuteWebAsciidocProcessor {
+
+    private static final String FEATURE = "qute-web-asciidoc-light";
+
+    @BuildStep
+    FeatureBuildItem feature() {
+        return new FeatureBuildItem(FEATURE);
+    }
+
+    @BuildStep
+    void process(BuildProducer<AdditionalBeanBuildItem> additionalBeans) {
+        additionalBeans.produce(new AdditionalBeanBuildItem(AsciidocLightSectionHelperFactory.class));
+    }
+
+    @BuildStep
+    void registerOpenPdfForReflection(BuildProducer<ReflectiveClassBuildItem> reflectiveClass,
+            CombinedIndexBuildItem combinedIndex) {
+        final List<String> classNames = new ArrayList<>();
+        classNames.add(io.yupiik.tools.ascii2svg.Svg.class.getName());
+
+        reflectiveClass.produce(
+                ReflectiveClassBuildItem.builder(classNames.toArray(new String[0])).methods().fields().build());
+    }
+
+}
