+++ asciidoc/asciidoc-light/integration-tests/src/main/java/io/quarkiverse/qute/web/it/MyBean.java
