+++ asciidoc/asciidoc-light/integration-tests/src/main/resources/templates/pub/hello.html
