+++ asciidoc/asciidoc-light/integration-tests/src/test/java/io/quarkiverse/qute/web/it/QuteWebAsciidocLightIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.qute.web.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class QuteWebAsciidocLightIT extends QuteWebAsciidocLightTest {
+}
