+++ asciidoc/asciidoctorj/integration-tests/src/test/java/io/quarkiverse/qute/web/it/QuteWebAsciidoctorJIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.qute.web.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class QuteWebAsciidoctorJIT extends QuteWebAsciidoctorJTest {
+}
