--- asciidoc/integration-tests/src/test/java/io/quarkiverse/qute/web/it/QuteWebAsciiDocIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.qute.web.it;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class QuteWebAsciiDocIT extends QuteWebAsciiDocTest {
-}
