--- asciidoc/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,14 +1,0 @@
-name: Qute Web Asciidoc
-description: A Qute Web extension designed to simplify the conversion of AsciiDoc to HTML by providing a set of AsciiDoc tags.
-metadata:
-  keywords:
-    - qute-web-asciidoc
-    - roq
-    - qute
-    - website
-    - templates
-    - asciidoc
-  guide: https://quarkiverse.github.io/quarkiverse-docs/quarkus-qute-web/dev/index.html
-  categories:
-    - "miscellaneous"
-  status: "preview"
