+++ markdown/runtime/src/main/java/io/quarkiverse/qute/web/markdown/runtime/MarkdownSectionHelperFactory.java
@@ -9,6 +9,7 @@ import io.quarkus.arc.Arc;
 import io.quarkus.arc.impl.LazyValue;
 import io.quarkus.qute.CompletedStage;
 import io.quarkus.qute.EngineConfiguration;
+import io.quarkus.qute.RawString;
 import io.quarkus.qute.ResultNode;
 import io.quarkus.qute.SectionHelper;
 import io.quarkus.qute.SectionHelperFactory;
@@ -47,8 +48,8 @@ public class MarkdownSectionHelperFactory
             () -> Arc.container().instance(MdConverter.class).get());
 
     @TemplateExtension(matchNames = { "markdownify", "mdToHtml" })
-    static String convertToMarkdown(String text, String ignoredName) {
-        return CONVERTER.get().html(text);
+    static RawString convertToMarkdown(String text, String ignoredName) {
+        return new RawString(CONVERTER.get().html(text));
     }
 
     public static class MarkdownSectionHelper implements SectionHelper {
