+++ pom.xml
@@ -30,8 +30,10 @@
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <quarkus.version>3.15.1</quarkus.version>
     <commonmark.version>0.24.0</commonmark.version>
-    <asciidoc.java.version>1.2.5</asciidoc.java.version>
+    <asciidoc.light.version>1.2.5</asciidoc.light.version>
+    <asciidoctorj.version>3.0.0</asciidoctorj.version>
     <assertj.version>3.26.3</assertj.version>
+    <asciidoctorj-diagram.version>2.3.1</asciidoctorj-diagram.version>
   </properties>
   <dependencyManagement>
     <dependencies>
